// examples/demo-machine/client/src/generated/variables.ts
export const Commands = {
  CatalogRevision: Opcua.variable({
    nodeId: NodeIds.Commands.CatalogRevision,
    codec: Opcua.schema(Schema.Number),
  }),
  Status: Opcua.variable({
    nodeId: NodeIds.Commands.Status,
    codec: Structures.CommandStatusBuffer,
  }),
  SubmitRequest: Opcua.variable({
    nodeId: NodeIds.Commands.SubmitRequest,
    codec: Structures.GlobalCommandSubmitRequest,
    access: "readWrite",
  }),
} as const;